Fix maskedvbyte and varintgb buffer initialization when encoding
Previous this commit, as thread_local is used to define the buffer, the first encode call initializes its size by using the current chunk size for any subsequent call. If in any of these calls, the current the output exceeds the previously defined buffer, then a segmentation fault will be thrown.